fix issues from Arnaud's review

+++ modules/core/include/Nystrom.h
@@ -131,7 +131,7 @@ private:
             MatrixType& U, VectorType& D) const {
         unsigned kernelDim = kernel->GetDimension();
 
-        double n = xs.size();
+        unsigned n = xs.size();
         MatrixTypeDoublePrecision K = MatrixTypeDoublePrecision::Zero(
                 n * kernelDim, n * kernelDim);
         for (unsigned i = 0; i < n; ++i) {
